fail if formatting is "required"https://github.com/talios/googleformatter-maven-plugin/issues/23
