+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>com.theoryinpractise</groupId>
     <artifactId>googleformatter-maven-plugin</artifactId>
-    <version>1.7.6-SNAPSHOT</version>
+    <version>1.8.0-SNAPSHOT</version>
     <packaging>maven-plugin</packaging>
 
     <name>googleformatter-maven-plugin Maven Mojo</name>
